SeLoversをLoversに移動
